chore(AP-3058): move non-pages out of pages directory
see: https://nextjs.org/docs/messages/prerender-error
